requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
alembic==1.13.1

# Validación y serialización
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2

# Documentación
drf-yasg==1.21.7

# Notificaciones
sendgrid==6.11.0

# Utilidades
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# Linting y formateo
black==24.1.1
flake8==7.0.0
pylint==3.0.3

# CORS
fastapi-cors==0.0.6